Aboutpage: autodiscover version & handle v3 vs v4
So v4 changed how the page was rendered a bit, and it now makes more
direct use of the api via javascript.  This is both obnoxious, in that
it broke the script, but ultimately much nicer since it means we can use
the same public api query to just snag the json and parse it FAR more
reliably.  That's now implemented

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>