Removing explicit nova list plugin
Nova stopped publishing the blocks to github and as a result the data
there is 'not good' and will semi-break the mastodon federation page.
As a note you should go and delete the rows with a blank domain and the
one that starts with `domain like 'More:%'` if you've borked the input.

ALSO as a result I'm now running validators on the domain name imported
to ensure it's a valid domain name in case this kind of thing happens
again.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>